Fix Drawing Angle offset to be counter-clockwise
... and use user-friendly Angle Selector to adjust it.

With this patch all three angle-related brush options are now converted to
"counter-clockwise" direction. That is, adding positive offset to Rotation,
Drawing Angle and Tilt Direction now rotates the brush counter-clockwise.

This makes all three options work somewhat consistent :)

Internally, Drawing Angle and Tilt Direction still save offsets in clockwise
manner, but they are inverted right before showing in the GUI.

+++ plugins/paintops/libpaintop/KisSensorData.cpp
@@ -129,7 +129,7 @@ void KisDrawingAngleSensorData::read(const QDomElement &e)
         fanCornersStep = e.attribute("fanCornersStep").toInt();
     }
     if (e.hasAttribute("angleOffset")) {
-        angleOffset = e.attribute("angleOffset").toInt();
+        angleOffset = e.attribute("angleOffset").toFloat();
     }
     if (e.hasAttribute("lockedAngleMode")) {
         lockedAngleMode = e.attribute("lockedAngleMode").toInt();
